conv3d/requirements.txt
pytorch-lightning
wandb
pillow
tqdm
numpy
pickle5